Return packet data copied to avoid race with Close
We need to copy packet data because as soon as ReadPacketData returns,
the ring buffer could be un-mapped by Close. If the caller of ReadPacketData
does not process packet data quickly enough and a call to Handle.Close()
interleaves, it could find itself reading from invalid memory segments.
If you are using go-pcap with google/gopacket, it is better to have packet
data copied by go-pcap and then enable NoCopy for PacketSource, i.e.:

  packetSource := gopacket.NewPacketSource(...)
  packetSource.NoCopy = true

Signed-off-by: Milan Lenco <milan@zededa.com>