Close handle without waiting for Poll to exit
After recent patches, Close() now waits for an ongoing ReadPacketData to
exit before unmapping memory and closing the socket. The problem is that
the reader may hang on the Poll(socket) call for as long as there is
no traffic matching the filter. I tried closing the socket as the first
action in Close(), but for AF_PACKET socket this didn't seem to have any
effect on an ongoing Poll, even if it was told to listen on POLLERR and
POLLNVAL events as well.

The solution is to add non-zero timeout to the Poll call and simply keep
repeating it for as long as reading is not canceled and there is no event.
A new state "polling" was introduced, which is set while the socket
is being polled by the reader. Close() is allowed to atomically change
"polling" to "cancelling". It is ensured that when polling is cancelled,
reader will not touch the ring buffer anymore and exit as soon as Poll
returns. This means that if Close is called during polling, it may
continue unmapping memory and closing socket without waiting for the
reader to exit and causing any delay for the caller.

Signed-off-by: Milan Lenco <milan@zededa.com>